requirements.txt
requests
ccxt>=4.2,<5
pandas>=2.2,<3
numpy>=1.26,<3
pyarrow>=15,<18
databricks-sdk>=0.16,<1
python-dotenv>=1.0,<2
